Refactoring setup_package stuff.
